PR feedback and refactoring
