+++ src/handlers.rs
@@ -1,2 +1,3 @@
+pub mod dapp_transaction_handler;
 pub mod utils;
 pub mod wallet_config_policy_update_handler;
