+++ src/processor.rs
@@ -1,7 +1,3 @@
-use bitvec::macros::internal::funty::Fundamental;
-use std::slice::Iter;
-use std::time::Duration;
-
 use solana_program::account_info::{next_account_info, AccountInfo};
 use solana_program::entrypoint::ProgramResult;
 use solana_program::hash::Hash;
@@ -20,9 +16,10 @@ use spl_token::instruction as spl_instruction;
 use spl_token::state::{Account as SPLAccount, Account};
 
 use crate::error::WalletError;
+use crate::handlers::dapp_transaction_handler;
 use crate::handlers::utils::{
     calculate_expires, collect_remaining_balance, get_clock_from_next_account,
-    next_program_account_info,
+    next_program_account_info, validate_balance_account_and_get_seed,
 };
 use crate::handlers::wallet_config_policy_update_handler;
 use crate::instruction::{BalanceAccountUpdate, ProgramInstruction, WalletUpdate};
@@ -186,7 +183,7 @@ impl Processor {
             ProgramInstruction::InitDAppTransaction {
                 ref account_guid_hash,
                 instructions,
-            } => Self::handle_init_dapp_transaction(
+            } => dapp_transaction_handler::init(
                 program_id,
                 accounts,
                 account_guid_hash,
@@ -196,7 +193,7 @@ impl Processor {
             ProgramInstruction::FinalizeDAppTransaction {
                 ref account_guid_hash,
                 ref instructions,
-            } => Self::handle_finalize_dapp_transaction(
+            } => dapp_transaction_handler::finalize(
                 program_id,
                 accounts,
                 account_guid_hash,
@@ -574,7 +571,7 @@ impl Processor {
         };
 
         if multisig_op.approved(&expected_params, &clock)? {
-            let bump_seed = Self::validate_balance_account_and_get_seed(
+            let bump_seed = validate_balance_account_and_get_seed(
                 source_account,
                 account_guid_hash,
                 program_id,
@@ -683,7 +680,7 @@ impl Processor {
             // it would be owned by the Token program). Since this is an attempt to wrap
             // SOL, it stands to reason they have some SOL in their account, so we assume
             // they have enough to create this account (if they don't, it will just fail)
-            let bump_seed = Self::validate_balance_account_and_get_seed(
+            let bump_seed = validate_balance_account_and_get_seed(
                 balance_account_info,
                 account_guid_hash,
                 program_id,
@@ -762,7 +759,7 @@ impl Processor {
         };
 
         if multisig_op.approved(&expected_params, &clock)? {
-            let bump_seed = Self::validate_balance_account_and_get_seed(
+            let bump_seed = validate_balance_account_and_get_seed(
                 balance_account_info,
                 account_guid_hash,
                 program_id,
@@ -948,195 +945,6 @@ impl Processor {
         Ok(())
     }
 
-    fn handle_init_dapp_transaction(
-        program_id: &Pubkey,
-        accounts: &[AccountInfo],
-        account_guid_hash: &BalanceAccountGuidHash,
-        instructions: Vec<Instruction>,
-    ) -> ProgramResult {
-        let accounts_iter = &mut accounts.iter();
-        let multisig_op_account_info = Self::next_program_account_info(accounts_iter, program_id)?;
-        let wallet_account_info = Self::next_program_account_info(accounts_iter, program_id)?;
-        let initiator_account_info = next_account_info(accounts_iter)?;
-        let clock = Self::get_clock_from_next_account(accounts_iter)?;
-
-        let wallet = Wallet::unpack(&wallet_account_info.data.borrow())?;
-        let balance_account = wallet.get_balance_account(account_guid_hash)?;
-
-        wallet.validate_transfer_initiator(balance_account, initiator_account_info)?;
-
-        let mut multisig_op =
-            MultisigOp::unpack_unchecked(&multisig_op_account_info.data.borrow())?;
-        multisig_op.init(
-            wallet.get_transfer_approvers_keys(balance_account),
-            1,
-            clock.unix_timestamp,
-            Self::calculate_expires(
-                clock.unix_timestamp,
-                balance_account.approval_timeout_for_transfer,
-            )?,
-            MultisigOpParams::DAppTransaction {
-                wallet_address: *wallet_account_info.key,
-                account_guid_hash: *account_guid_hash,
-                instructions,
-            },
-        )?;
-        MultisigOp::pack(multisig_op, &mut multisig_op_account_info.data.borrow_mut())?;
-        Ok(())
-    }
-
-    fn account_balances(accounts: &[AccountInfo]) -> Vec<u64> {
-        accounts.iter().map(|a| a.lamports()).collect()
-    }
-
-    fn spl_balances(accounts: &[AccountInfo]) -> Vec<SplBalance> {
-        accounts
-            .iter()
-            .filter_map(|a| {
-                if *a.owner == spl_token::id() {
-                    if let Some(account_data) = SPLAccount::unpack(&a.data.borrow()).ok() {
-                        Some(SplBalance {
-                            account: *a.key,
-                            token_mint: account_data.mint,
-                            balance: account_data.amount,
-                        })
-                    } else {
-                        None
-                    }
-                } else {
-                    None
-                }
-            })
-            .collect()
-    }
-
-    fn balance_changes_from_simulation(
-        starting_balances: Vec<u64>,
-        starting_spl_balances: Vec<SplBalance>,
-        accounts: &[AccountInfo],
-    ) -> String {
-        // compute and log just the changes to minimize compute budget spend
-        let ending_balances = Self::account_balances(accounts);
-        let mut balance_changes: Vec<(u8, char, u64)> = Vec::new();
-        for i in 0..starting_balances.len() {
-            if ending_balances[i] > starting_balances[i] {
-                balance_changes.push((i as u8, '+', ending_balances[i] - starting_balances[i]));
-            } else if ending_balances[i] < starting_balances[i] {
-                balance_changes.push((i as u8, '-', starting_balances[i] - ending_balances[i]));
-            }
-        }
-        let ending_spl_balances = Self::spl_balances(accounts);
-        let mut spl_balance_changes: Vec<(u8, char, u64)> = Vec::new();
-        for end in ending_spl_balances.into_iter() {
-            let starting_balance = starting_spl_balances
-                .iter()
-                .find(|start| start.account == end.account && start.token_mint == end.token_mint)
-                .map(|start| start.balance)
-                .unwrap_or(0);
-            if end.balance > starting_balance {
-                spl_balance_changes.push((
-                    accounts
-                        .iter()
-                        .position(|a| *a.key == end.account)
-                        .unwrap()
-                        .as_u8(),
-                    '+',
-                    end.balance - starting_balance,
-                ));
-            } else if starting_balance < end.balance {
-                spl_balance_changes.push((
-                    accounts
-                        .iter()
-                        .position(|a| *a.key == end.account)
-                        .unwrap()
-                        .as_u8(),
-                    '-',
-                    starting_balance - end.balance,
-                ));
-            }
-        }
-        format!(
-            "Simulation balance changes: {:?} {:?}",
-            balance_changes, spl_balance_changes
-        )
-    }
-
-    fn handle_finalize_dapp_transaction(
-        program_id: &Pubkey,
-        accounts: &[AccountInfo],
-        account_guid_hash: &BalanceAccountGuidHash,
-        instructions: &Vec<Instruction>,
-    ) -> ProgramResult {
-        let accounts_iter = &mut accounts.iter();
-        let multisig_op_account_info = Self::next_program_account_info(accounts_iter, program_id)?;
-        let wallet_account_info = Self::next_program_account_info(accounts_iter, program_id)?;
-        let balance_account = next_account_info(accounts_iter)?;
-        let rent_collector_account_info = next_account_info(accounts_iter)?;
-        let clock = Self::get_clock_from_next_account(accounts_iter)?;
-
-        if !rent_collector_account_info.is_signer {
-            return Err(ProgramError::MissingRequiredSignature);
-        }
-
-        let multisig_op = MultisigOp::unpack(&multisig_op_account_info.data.borrow())?;
-
-        let expected_params = MultisigOpParams::DAppTransaction {
-            wallet_address: *wallet_account_info.key,
-            account_guid_hash: *account_guid_hash,
-            instructions: instructions.clone(),
-        };
-
-        let is_approved = match multisig_op.approved(&expected_params, &clock) {
-            Ok(a) => a,
-            Err(e) => {
-                msg!("Approval failed: {:?}", e);
-                false
-            }
-        };
-
-        let bump_seed = Self::validate_balance_account_and_get_seed(
-            balance_account,
-            account_guid_hash,
-            program_id,
-        )?;
-
-        let starting_balances: Vec<u64> = if is_approved {
-            Vec::new()
-        } else {
-            Self::account_balances(accounts)
-        };
-
-        let starting_spl_balances: Vec<SplBalance> = if is_approved {
-            Vec::new()
-        } else {
-            Self::spl_balances(accounts)
-        };
-
-        for instruction in instructions.iter() {
-            invoke_signed(
-                &instruction,
-                &accounts,
-                &[&[&account_guid_hash.to_bytes(), &[bump_seed]]],
-            )?;
-        }
-
-        if is_approved {
-            Self::collect_remaining_balance(
-                &multisig_op_account_info,
-                &rent_collector_account_info,
-            )?;
-
-            Ok(())
-        } else {
-            msg!(&Self::balance_changes_from_simulation(
-                starting_balances,
-                starting_spl_balances,
-                accounts,
-            ));
-            Err(WalletError::SimulationFinished.into())
-        }
-    }
-
     fn collect_remaining_balance(from: &AccountInfo, to: &AccountInfo) -> ProgramResult {
         // this moves the lamports back to the fee payer.
         **to.lamports.borrow_mut() = to
@@ -1172,24 +980,4 @@ impl Processor {
             &[&[&account_guid_hash.to_bytes(), &[bump_seed]]],
         )
     }
-
-    fn validate_balance_account_and_get_seed(
-        balance_account: &AccountInfo,
-        account_guid_hash: &BalanceAccountGuidHash,
-        program_id: &Pubkey,
-    ) -> Result<u8, ProgramError> {
-        let (account_pda, bump_seed) =
-            Pubkey::find_program_address(&[&account_guid_hash.to_bytes()], program_id);
-        if &account_pda != balance_account.key {
-            Err(WalletError::InvalidSourceAccount.into())
-        } else {
-            Ok(bump_seed)
-        }
-    }
-}
-
-struct SplBalance {
-    account: Pubkey,
-    token_mint: Pubkey,
-    balance: u64,
 }
