+++ tests/common/mod.rs
@@ -1,1 +1,2 @@
+pub mod instructions;
 pub mod utils;
